gameobj-data: fix issues with Red Forest skins
Treekin warriors yield a "some blood-stained bark" as skins. Unlike
other skins, this skin has a non-countable determiner (some) which
is included as part of the name, unlike the countable determiners a/an
which the game presents as outside the name. These also have a slightly
different bundled name (bundle of blood-stained barks) which needed to
be accounted for.

The compiled version of gameobj-data.xml was also missing the
"pure white feather" yielded by winged vipers.