Orb outcome from 1004 overrides other outcomes
When using 1004 on a gem, the gem damage messaging (i.e. improvement or
strain) comes before the messaging for a gem purified to orb status.
This updates the script to continue processing the 1004 result after the
gem damage messaging until sees the orb messaging, the generic non-orb
messaging from 1004, or another result that cannot result in an orb gem.