Improving treim 1030 reaction time
Treim 1030 has the unfortunate side effect of issuing several commands. This prevents getting hits in on many fast-dying waves. As such, I've modified the routine to send a single command that handles the casting: `incant 1030 open`

Based on verified bardic research, this command uses the RENEW 1030 cost where appropriate and is open-targeted as desired.