bigshot: track obvious hiding msg separately from ambusher_here
$ambusher_here is reset during variables_reset, which is called after a
character walks into a new room, which is also after the server_string
has already been processed (so the subsequent variables_reset then
clears $ambusher_here immediately after the flag is set).

This switches to using a separate variable and resets it in a different
spot to work around this issue.