Update tfish.lic
I like the change to generalized weight instead of using the variable. My OCD is kicking in though as cast_counter routine was previously no weight, depths, bottom and it was changed to depths, bottom, no weight. So modified that and added a version change at the top. Thanks for the idea and hope you don't mind my OCD putting it back to no weights, depths, bottom. The only issue I can see later on is if they release weights using a different noun, but can cross that bridge when Simu decides to use alternative nouns.