Fix for ignoring players in first room of a hunt
When starting a hunt, bigshot ignores characters in the first room as it checked no_players_hunt() which only checks $ambusher_here and $obvious_hiding_player which are both defaulted to false at the start of a hunt. Changed it to check no_players(). The only concern I have is in busy areas with a lot of traffic could cause critters to be skipped from disks trailing behind their masters.