[bigshot.lic] fixed shield bash for semi's
Changed 1040 to have a checkbox. When enabled will check self and group members for status ailments, and shout 1040 to attempt to clear them. 5 second timeout to keep from multiple casts back to back. This may need to be adjusted. 1040 seems to be RNG without heavy lores. Will need to find the balance between effective and blowing all your mana, or maybe this is just left up to the user.
Pull players to feet will no longer pull kneeling players.