[bigshot.lic] commented out ammo gathering
Putting your ammo in the ammo tab triggers the ammo gather portion of the code. It also appears to be used in the bless portion. I commented out all the calls to gather ammo as well as the call to remove an arrow to fire since this is all no longer needed with the updates to archery.
The only use case I can see for leaving the pull 1 arrow out part is if you want to use a specific ammo for fire and your ready ammo for psm skills. If this issue comes up we can seek a different solution.