[bigshot.lic] updates to room claim
Entering a room pulls a buffer, buffer is checked for Also here: indicating another player was there first and claiming the room. If this returns nothing it then checks against checkpcs to verify incase buffer was over run by behind the scenes commands. Removes players if they have a claim and walk out of the room. So having a wait before moving time can be beneficial when you end up "chasing" someone.