[bigshot.lic] updates to room claim
Change room claim from relying on checkpcs to first pull a buffer when entering the room and looking for Also here:. If nothing is found, then it falls back to checking checkpcs as a backup.  If someone has a claim and they walk out of the room, their claim is removed.