[bigshot.lic]
Removed a piece of the regex to remove claim. It was some how matching on sorcerer prep messaging even though the \. should have prevented it.
Not sure what that specific piece of regex was suppose to match, likely some custom movement messaging.
Seperated the two regex's remaining to prevent weirdness?