[Overwatch.lic] Initial creation (#4)
* [Overwatch.lic] Initial creation

A script that watches for creatures to be revealed from hiding. Adds them to GameObj.npc and XMLData.current_target_ids if they are not present.

Also attempts to determine when there is a creature currently hidden in the room, setting a variable with the room number. This variable is cleared when a hidden target is revealed or you change rooms. 
OverWatch.room_has_hiders  will return the room number for hidden creatures or nil for no known hidden creatures.

* Update overwatch.lic

Rubocop

* Update overwatch.lic

O's suggestions.
Corrections and adding OverWatch.hiders?