[Targetwindow.lic]v1.0
Put the status in front of the name. Some OSA creatures have really long names. Would rather the name be cut off when the window is at it's smallest width than for the status to be cut off.

Put a little more space between lines, making it less easy to misclick.