[Ebounty.lic] v1.1.35 implement nomagic rooms
It's probably far and few between, but at least in the atoll there is a forage bounty for seaweed that has you on the log which is nomagic. If using 506 this causes you to cast and fail, cast and fail, then forage. Decided to see about implementing the nomagic tag that Xanlin added to the mapdb. Feel free to change any of this as I'm probably not considering everything here.

Adding a check for `Map.current.tags.include?("meta:nomagic")` when casting spells.