Improved ready_to_rest? logic to reduce misleading messages
There are a few places in the script that set $bigshot_should_rest to true. For example, in the cmd_spell function. If a spell is unaffordable, it sets $bigshot_should_rest to true and provides a reason.

However in the ready_to_rest? function, it immediately sets $rest_reason to nil, erasing any indicator of a pre-existing rest reason. This adds a new rest_condition and rest_action that simply captures this pre-existing reason. 

It also moves should_rest to the end so it acts as a "default" reason, allowing more detailed reasons to be processed first