[ecaster.lic] add option to hide
;ec hide <spell_number>  - will add/remove spell to @@hide list.
;ec set hide true/false  -  will enter a single hide command after you cast.

The question is where to place the hide? Before or after the stance change?
I like it before the stance change because that allows the use of a `put` to provide a .1-.2 second delay between casting and hiding. This would require a short sleep and waitrt before the stance change or you get a wait 1 second message from the stance command.
It could go after the stance command. This takes 2x-3x the time to hide as putting it before the stance. We're still talking something like .5-.7s to be hidden.

This actually makes no difference to me personally as the only spell I use that uses a stance is 615 swarm. So I get the instant hide before or after the stance.
Who is actually going to use this though? Rangers. Rogues for the occasional ewave?  Nobody other than me since it hasn't been added yet?


I think it should be before, but I'm an excellent hider. Failing the hide could leave you in an offensive stance exposed for 1-3 seconds.
It doesn't actually effect me, because I don't stance for my spells and the stance section just gets skipped over anyways.
Is anyone actually casting/hiding other than rangers/rogues?

I placed the hide after the cast, before the stance. I understand some people may prefer it after the stance?
It doesn't really affect me as none of the spells I use need me to stance.
If you fail the hide, you have 1-3s of hard RT.
If you do after the stance, are you going to be stuck waiting until it gets you back into