[SpellWindows.lic] incorporate additional windows
Incorporated buff_tracker functionality of a customizable list of spells to list if they're not active.
Incorporated targetwindow functionality and expanded to list players and group members.

These are optional windows that are disabled by default. Incorporating them into a single script and reducing threads is a plus.

This video shows a character with one group member running to town and back. The top screen is a third character that shows the actual arrival time vs the experienced arrival time. Neither character have any scripts running to start. The second run is group leader with spellwindows active. Third run is both characters with spellwindows active.
https://www.youtube.com/watch?v=DoUZUshtOk4