feat(eforgery.lic):  Add support for using notes (#1873)
v1.2.0 (2025-03-22)
Will uses notes on you instead of running to the bank every action.
Thanks to Deyesh as I have heavily borrowed from his work for this.
    Modified the trashing routine to use the "trash" verb
Modified the trash barrel location for Ta'Vaalor to be closer to the
forge
    Some cosmetic formatting updates for readability.
Added the important read this info at the top and a script pause for the
first run.
    Added debugging code.