Update kswole.lic shield mind -> shield mind forcert
Whirlin suggested that shield mind could be executed during rt and he was correct after testing.  Given the nature of the ability, we would always want to execute the command during round time to avoid taking a big hit from a spell.

Removed pausing a script as it is no longer needed.
Removed waitrt? and waitcastrt? as it is no longer needed.