restructured to better handle destinations
removed checking for Char.che validity (it should always have a value now in 5.12.2).

updated to better handle that there could be a mix of locker and entrance_locker tags, and to try to remove misleading entrance_locker tags that may prevent efficient access to an actual locker because it is technically a shorter distance.