fix(gs): Killtracker fix timestamp validation
Corrected timestamp validation to ignore timestamps below 100,000.

If someone did not log with timestamps, and they used the logsearch script to populate killtracker data, it would fill in the timestamps with the oversall ascension creature search # it has. This resulted in "timestamps" that are way out of an acceptable range. Corrected by ignoring any timestamp below 100,000..

The validation doesn't effect the script actually working, it's more of a check to alert you of possible corrupted data before migrating to the new format.